+++ plugins/postcss-text-decoration-shorthand/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Text Decoration Shorthand
 
+### Unreleased (patch)
+
+- Reduce redundant fallbacks for both `text-decoration` and `-webkit-text-decoration`
+
 ### 5.0.1
 
 _January 25, 2026_